requirements.txt
aiohttp==3.7.4.post0
anyio==3.2.1
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.6.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.2
click==8.0.1
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.66.0
fsspec==2021.7.0
gcsfs==2021.7.0
google-api-core==1.31.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
h11==0.12.0
idna==3.2
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pylint==2.9.3
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
wrapt==1.12.1
yarl==1.6.3
